+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 * [#1923](https://github.com/bbatsov/rubocop/issues/1923): Handle properly `for` without body in `Style/Next`. ([@bbatsov][])
 * [#1901](https://github.com/bbatsov/rubocop/issues/1901): Do not auto correct comments that are missing a note. ([@rrosenblum][])
 * [#1926](https://github.com/bbatsov/rubocop/issues/1926): Fix crash in `Style/AlignHash` when correcting a hash with a splat in it. ([@rrosenblum][])
+* [#1935](https://github.com/bbatsov/rubocop/issues/1935): Allow `Symbol#to_proc` blocks in Performance/Size. ([@m1foley][])
 
 ## 0.31.0 (05/05/2015)
 
@@ -1437,3 +1438,4 @@
 [@RGBD]: https://github.com/RGBD
 [@panthomakos]: https://github.com/panthomakos
 [@matugm]: https://github.com/matugm
+[@m1foley]: https://github.com/m1foley